requirements.txt
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
flask-cors==4.0.0
pymongo==4.3.3
weaviate-client==3.19.1
openai==0.27.6
langchain==0.0.173
typing_extensions==4.5.0
tiktoken==0.4.0
Werkzeug==2.3.3